Extract `ValueExtractingTensorGenerator` from `Constant`.
`ZerosLike` and `ConvertToTensor` `extends Constant` to borrow its
value-argument shape/dtype inference, and `ElementWiseOperation` once
chained through it too (wala/ML#514). `zeros_like` and
`convert_to_tensor` are not kinds of `constant`; the shared abstraction
is "a tensor whose shape and dtype are extracted from a value/input
argument."

Extract that into an abstract `ValueExtractingTensorGenerator` (value at
position 0, dtype at position 1, no explicit shape by default).
`Constant` extends it and adds an explicit `shape` argument; `ZerosLike`
extends it and renames the value argument to `input`; `ConvertToTensor`
extends it and adds a `dtype_hint` fallback. `SparseFromDense` continues
to extend `ConvertToTensor`. Behavior-preserving; full suite green (970).
