+++ com.ibm.wala.cast.python.ml/source/com/ibm/wala/cast/python/ml/client/Constant.java
@@ -1,23 +1,17 @@
 package com.ibm.wala.cast.python.ml.client;
 
-import com.ibm.wala.cast.python.ml.types.TensorFlowTypes.DType;
-import com.ibm.wala.cast.python.ml.types.TensorType.Dimension;
-import com.ibm.wala.ipa.callgraph.propagation.InstanceKey;
 import com.ibm.wala.ipa.callgraph.propagation.PointsToSetVariable;
-import com.ibm.wala.ipa.callgraph.propagation.PropagationCallGraphBuilder;
-import com.ibm.wala.util.intset.OrdinalSet;
-import java.util.EnumSet;
-import java.util.List;
 import java.util.Locale;
-import java.util.Set;
 
 /**
- * Represents a call to the <code>constant()</code> function in TensorFlow.
+ * Represents a call to the <code>constant()</code> function in TensorFlow. The value-argument
+ * shape/dtype inference lives in {@link ValueExtractingTensorGenerator}; {@code tf.constant} adds
+ * an optional explicit {@code shape} argument.
  *
  * @see <a href="https://www.tensorflow.org/api_docs/python/tf/constant">constant()</a>.
  * @author <a href="mailto:khatchad@hunter.cuny.edu">Raffi Khatchadourian</a>
  */
-public class Constant extends TensorGenerator {
+public class Constant extends ValueExtractingTensorGenerator {
 
   protected enum Parameters {
     VALUE,
@@ -40,51 +34,6 @@ public class Constant extends TensorGenerator {
   }
 
   @Override
-  protected Set<List<Dimension<?>>> getDefaultShapes(PropagationCallGraphBuilder builder) {
-    // If the shape argument is not specified, then the shape is inferred from the shape of value.
-    OrdinalSet<InstanceKey> pointsToSet =
-        this.getArgumentPointsToSet(
-            builder, this.getValueParameterPosition(), this.getValueParameterName());
-    Set<List<Dimension<?>>> shapes = this.getShapesOfValue(builder, pointsToSet);
-    // An empty result means the shape could not be inferred — treat as unknown (null / ⊤).
-    // `getShapesOfValue` documents that it returns `null` (not an empty set) when the
-    // value's points-to set is empty; honor that contract instead of NPE-ing.
-    return (shapes == null || shapes.isEmpty()) ? null : shapes;
-  }
-
-  /**
-   * {@inheritDoc}
-   *
-   * <p>If the <code>dtype</code> argument is not specified, then the type is inferred from the type
-   * of value.
-   */
-  @Override
-  protected Set<DType> getDefaultDTypes(PropagationCallGraphBuilder builder) {
-    OrdinalSet<InstanceKey> pointsToSet =
-        this.getArgumentPointsToSet(
-            builder, this.getValueParameterPosition(), this.getValueParameterName());
-    Set<DType> dTypes = this.getDTypesOfValue(builder, pointsToSet);
-    // An empty result means the dtype could not be inferred — treat as unknown
-    // (⊤), mirroring the ⊤ treatment in getDefaultShapes. A `tf.constant` is
-    // always a tensor; we just cannot pin down its dtype.
-    // `getDTypesOfValue` documents that it returns `null` (not an empty set) when
-    // the value's points-to set is empty; honor that contract instead of NPE-ing.
-    return (dTypes == null || dTypes.isEmpty()) ? EnumSet.of(DType.UNKNOWN) : dTypes;
-  }
-
-  protected int getValueArgumentValueNumber() {
-    return this.getArgumentValueNumber(this.getValueParameterPosition());
-  }
-
-  protected int getValueParameterPosition() {
-    return Parameters.VALUE.getIndex();
-  }
-
-  protected String getValueParameterName() {
-    return Parameters.VALUE.getName();
-  }
-
-  @Override
   protected int getShapeParameterPosition() {
     return Parameters.SHAPE.getIndex();
   }
@@ -92,15 +41,5 @@ public class Constant extends TensorGenerator {
   @Override
   protected String getShapeParameterName() {
     return Parameters.SHAPE.getName();
-  }
-
-  @Override
-  protected int getDTypeParameterPosition() {
-    return Parameters.DTYPE.getIndex();
-  }
-
-  @Override
-  protected String getDTypeParameterName() {
-    return Parameters.DTYPE.getName();
   }
 }
